backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
mysqlclient==2.2.7
django-cors-headers==4.7.0